requirements.txt
streamlit>=1.25.0
pandas
numpy
faiss-cpu
sentence-transformers
transformers
cohere
